dvpn: Delete the correct PEER LSA attr when destroying a conf_listen_entry.
When a listen entry which has a client connected to it is destroyed, we
would attempt to delete from the local LSA a PEER attribute with a key
matching the PeerFingerprint of the listen entry, but this is incorrect
if the client has connected using a role certificate.  Use the connected
peer's ID instead of the listen entry's configured PeerFingerprint.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>