tconn: Fix gnutls_record_recv() GNUTLS_E_AGAIN return handling.
As it turns out, gnutls_record_recv() returning GNUTLS_E_AGAIN doesn't
necessarily mean that the underlying socket's receive buffer is empty.
We did assume that that was the case, and that caused the state
validation logic in verify_state() to bail out sometimes.

Fix this by repeatedly calling gnutls_record_recv() when it returns
GNUTLS_E_AGAIN as long as we still have data in the receive buffer,
and add a detailed description to tconn_do_record_recv() explaining
this issue and how we work around it.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>